Move source files to folder
