+++ nrrd/__init__.py
@@ -1,0 +1,4 @@
+from nrrd.reader import read, read_data, read_header
+from nrrd.writer import write
+
+__all__ = ['read', 'read_data', 'read_header', 'write']
